+++ tests/schema/test_schema.py
@@ -18,7 +18,7 @@ from git import Head
 from mock import patch
 
 from autotransform.batcher.base import Batch
-from autotransform.batcher.single import SingleBatcher, SingleBatcherParams
+from autotransform.batcher.single import SingleBatcher
 from autotransform.change.base import Change
 from autotransform.filter.regex import RegexFilter
 from autotransform.input.directory import DirectoryInput
@@ -40,7 +40,7 @@ def get_sample_schema() -> AutoTransformSchema:
     repo_root = str(pathlib.Path(__file__).parent.parent.parent.resolve()).replace("\\", "/")
     return AutoTransformSchema(
         DirectoryInput({"path": repo_root}),
-        SingleBatcher(SingleBatcherParams(title=EXPECTED_TITLE, metadata=EXPECTED_METADATA)),
+        SingleBatcher(title=EXPECTED_TITLE, metadata=EXPECTED_METADATA),
         RegexTransformer({"pattern": "input", "replacement": "inputsource"}),
         SchemaConfig("Sample", owners=["foo", "bar"]),
         filters=[RegexFilter({"pattern": ".*\\.py$"})],
@@ -308,12 +308,7 @@ def test_json_decoding(_mocked_checkout):
     ), "Inputs do not have the same params"
 
     # Check batcher
-    actual_batcher = actual_schema.get_batcher()
-    expected_batcher = expected_schema.get_batcher()
-    assert type(actual_batcher) is type(expected_batcher), "Batchers are not the same"
-    assert (
-        actual_batcher.get_params() == expected_batcher.get_params()
-    ), "Batchers do not have the same params"
+    assert actual_schema.get_batcher() == expected_schema.get_batcher()
 
     # Check transformer
     actual_transformer = actual_schema.get_transformer()